+++ packages/web/src/header/Header.3.tsx
@@ -12,7 +12,7 @@ import Octocat from 'src/icons/Octocat'
 import LogoDarkBg from 'src/logos/LogoDarkBg'
 import LogoLightBg from 'src/logos/LogoLightBg'
 import Button, { BTN } from 'src/shared/Button.3'
-import Footer from 'src/shared/Footer.3'
+import Footer from 'src/shared/Footer'
 import Link from 'src/shared/Link'
 import menu, { CeloLinks } from 'src/shared/menu-items'
 import OvalCoin from 'src/shared/OvalCoin'
