+++ packages/web/src/shared/Footer.tsx
